Call `doit()` on Subs objects in diffgeom.
Due to employing the chain rule many expression in diffgeom end up in the form
Subs(expr, x, y) where substituting x for y is trivially correct, however it is
not done. Calling `doit()` by default eleviates this problem, however it may
cause problems if the user mixes in other expressions that he does not want
evaluated. A better solution will be employing a custom `doit` routine that
applies only to `Subs` instances.