Preprocess strings containing relational operators in sympify
Now this works:

```
In [1]: sympify(parse_relationals('a<b'))
Out[1]: a < b

In [2]: sympify(parse_relationals('a==b'))
Out[2]: a = b

In [3]: sympify(parse_relationals('sin(a)==b'))
Out[3]: sin(a) = b
```
etc...

Fixes issue 1625

Due to how different this behavior is from that of python's `eval` I am not
adding it directly to `sympify`. When we implement more parsing functions we
will be able to build a better self-contained parser.