Reset PYTHONHASHSEED after running the tests or doctests
This way, it will be as it was for the outer Python process (the one that
creates the subprocess).  Without this, running more than one test (e.g., the
tests and then the doctests) would not work after the first one, because the
later ones would think that the seed was already set and not spawn a
subprocess.

Note that as a result of this, each test/doctest will have a separate hash
seed.