DiagramGrid: Add a test for a strange diagram.
This test attempts to use DiagramGrid to lay out a diagram which
resulted from a typo in an attempt to lay out the  the five lemma.
DiagramGrid currently fails to lay it out properly because of
the fact that en route it encounters a situation where none of the
remaining triangles can be welded to the already constructed structure.