Docstrings, tests, refactoring
Wrote docstrings and tests for the new functionality,
and did some refactoring in the new functions.

Also, made possible the interaction with the
deterministic version of the Schreier-Sims algorithm
by introducing some new attributes.

Reorganized the util.py file containing helper functions for
the combinatorics module.

Altered _strip in sympy/combinatorics/util.py
so that it accepts basic transversals as an argument, instead of
basic stabilizers, so that finding transversal elements is done
in constant time. Altered _distribute_gens_by_base in the same
file slightly.