Add make html-errors for the doc directory
This just adds the -W flag to sphinx-build, which treats all warnings as
errors.  This is useful for example with sympy-bot, where one can do

./sympy-bot review 1400 -t "-V; cd doc; make html-errors"

and the bot will report passing status iff the docs build without any errors
or warnings (the -V is necessary because the bot passes the -t argument to
python).