DiagramGrid._triangle_key: Remove the necessity of objects to have names.
This method used to produced a key for sorting the triangles which was
dependent on objects having names.  This would fail for groups, where
objects are FiniteSet's.

Note that this method currently sorts the set of objects in order to
avoid the problem of the sort key of FiniteSet being different in
dependence of the order of the elements in it.