XypicDiagramDrawer: Decouple the analysis of morphisms from typesetting.
This commit makes the analysis of morphisms in view of producing the
corresponding string representation a separate stage, which happens
before the actual generation of Xy-pic code.  This allows traversing the
morphisms in any order, as well as modifying the properties of some
morphisms after they have already been analysed.