3322: don't let -1 and I join in Mul._eval_subs
If -1 and I appear in a pattern, they will be combined when I's base and
exponent are given as -1 and 1/2. One way to modify sympy so this doesn't
occur is to change I's as_base_exp method to give (I, 1) but then expressions
like -(-1)**(3/4)*I won't combine properly, so we just watch for the negative
coefficient on self and pull it out before proceeding.