IntConst is no more detected as a constant by odesymp.
IntConst was introduced as a first step in solving issue 1739. It separates the
logic for constant simplification used in the single equation ode solver. This
permits the system of equations ode solver to work with constants without
worrying about a constant getting renumbered. In order to complete this change
the present commit removed the 'C' prefix of all IntConst instances. This stops
odesimp from modifying them.