randomization: Subs uses dummies as private entities
Subs no longer uses dummy variables to rewrite the arguments of the
instance since this doesn't help in making comparisons between different
instances. Instead, symbols whose name is the string form of the point
value being substituted is used instead, but these symbols are used to
store a private expr where the symbols in the expr are replaced with these
new symbols (stored in ._expr and used for the hash and for making
comparisons in __eq__).

e.g.

Subs(x, x, 0)
    .expr = x
    .variables = (x,)
    .point = (0,)
    ._expr = _0