Second "constant of integration" class for internal use.
Most of the internal routines in this module expect
IntegrationConstant('C1')==IntegrationConstant('C1'), thus we can not use
the IntConst class as it subclasses Dummy. All instances of this class are
changed to instances of IntConst at the end.

The need for this class arises because the output of dsolve may be used as
an input to another call to dsolve. Hence dsolve needs to be able to
distinguish newly generated constants from those that were alreadyresent
in the input."""

This solves issue 1739