XypicDiagramDrawer: Handle multiple morphisms between the same two objects.
XypicDiagramDrawer is now able to deal with the situation when there is
more than one morphism between the same two objects.  These morphisms
are drawn curved.  Nothing is done to decide which direction of curving
would be more efficient; this is thus an opportunity for further
optimisation.

The test this commit adds also shows that, because of multiple curved
morphisms, the diagram may become rather crowded.  This is easily
overcome by increasing row and column spacing of the diagram in the
Xy-pic output.