DiagramGrid._handle_groups: Add lay_out_group.
Previously, the part of _handle_groups that laid out each of the group
would consist of two largely similar halves: one that processed groups
supplied as an array of FiniteSet's, and the other one which processed
groups as a dictionary mapping groups to their hints.  This commit
factors out the common part into lay_out_group.