Remove Integral._eval_expand_basic
The only difference between this and Integral.expand() was that the former did
not expand the limits, which was a subtle difference at best.  The 'basic'
hint remains in the API for any object that wants expansion to happen
automatically without having to add hint=True to the expand call, but with
this change, no object in SymPy itself uses the 'basic' hint.