Factor _print_matrix_contents out of _print_MatrixBase.
This commit adds to the string pretty printer the method
_print_matrix_contents which only carries that part of the previous
version of _print_MatrixBase which refers to printing the grid.
_print_MatrixBase now only puts this grid in brackets.

The motivation behind this commit is that grid printing is useful in
itself and thus should be available as a separate method.  See the next
commit for an example of such usage.