Pow.__new__ and (Mul|Add).flatten edits
Since e==0 is already teste at the outset
this condition can be removed from the test for S.NaN

Adding special 'quick exit' handling of the exponent 0
(or factor of 1 in Mul.flatten) doesn't lead to as significant
speedups as simply handling those special cases in the calling
routines. Apparently the overhead to calling the routine is
much greater than the extra handling of those cases in-line.

A similar note was added to Add.flatten regarding the handling of 0
(which is analogous to a factor of 1 for Mul).