add additional test for issue 2644
This test guards against the assumption that only `term` need be searched
for Adds. If t.args is replaced with term.args, the second test will
fail while the existing test will pass because term=sqrt(x+y) and its
args are (x+y, 1/2) -- one of which is an Add -- so the expansion is done.