DiagramGrid._grow_pseudopod: Fix a very ugly mistake.
Initially I forgot to pick a certain element from the list of triangles
which could be attached and used the implicit value of the loop
variable, which would be the last element of the list.  This commit
fixes this mistake by explicitly picking the first triangle in the list.

This change has slightly changed the result for a test case; this commit
fixes the test as well.