New matrix method for finding particular solutions of A*x=b.
A routine for row elimination (including for singular matrices) was already
implemented inside a function in `solvers.py`. This implementation was refactored
in a separate function moved to `matrices.py`. This same implementation is now
used for finding particular solutions of matrix equations. It works for systems
which were not solvable with the other `Matrix.*solve()` methods.