randomization: combsimp
The result obtained from combsimp is dependent on the order of processing
in several places (as indicated by the use of sorted and _sorted_args).

Now, although the answer may not be optimal, it is canonical.

e.g., here are 2 equally valid answers that are obtained depending on
the order of processing:

>>> a
3**(3*x + 1/2)/(18*pi*gamma(3*x - 1))
>>> b
3**(3*x - 3/2)/(2*pi*gamma(3*x - 1))
>>> [w.subs(x,2).n() for w in (a,b)]
[0.930367511024274, 0.930367511024274]