edit: default_sort_key docstring; handle sets
The docstring explains why using the order hint may not
give the results one expects if using the default_sort_key
for sorting rather than canonical ordering.

The way around this would be
to create a new sort function that collects a set Expr items,
then assigns an order to each one as might be given if those
items were in an Add or Mul. The only catch is that 0 cannot appear in
an Add and 1 cannot appear in a Mul:

>>> Add(1/x, x ,1, 2, evaluate=0).as_ordered_terms('rev-lex')
[1/x, 2, 1, x]
>>> Add(1/x, x ,1, 2, evaluate=0).as_ordered_terms()
[x, 1, 2, 1/x]

>>> Mul(0, 1/x, x + 1, x**2, evaluate=0).as_ordered_factors()
[0, 1/x, x**2, x + 1]