Add -i, --int-to-Integer option to isympy
This makes isympy automatically wrap integer literals with Integer.  The
result is that things like 1/2 are computed as Rational(1, 2) instead of 0.5.
For example:

$./bin/isympy -i
IPython console for SymPy 0.7.1-git (Python 2.7.3-64-bit) (ground types:
python)

These commands were executed:
>>> from __future__ import division
>>> from sympy import *
>>> x, y, z, t = symbols('x y z t')
>>> k, m, n = symbols('k m n', integer=True)
>>> f, g, h = symbols('f g h', cls=Function)

Documentation can be found at http://www.sympy.org

In [1]: 1/2
Out[1]: 1/2

In [2]: x**(1/2)
Out[2]:
    ___
  ╲╱ x

You can still get a int literal by wrapping the integer with int()

In [3]: int(1)/int(2)
Out[3]: 0.5

Note that this requires IPython.

Also added -I, --interactive, which combines this option with -a,
--auto-symbols (which has been renamed from --auto).  Additional such
interactive modifications should be included in this command line argument.

Note that this had to be done by monkey-patching IPython's run_cell. IPython's
current API for preprocessing lines is not expressive enough to handle
this--in particular, it does not allow modifying multiple line inputs.  See
http://mail.scipy.org/pipermail/ipython-user/2012-August/010846.html and
https://github.com/ipython/ipython/issues/1491 for more information.

Thanks to http://stackoverflow.com/a/11735604/161801 for suggesting the
tokenize method of doing this.

This fixes issue 3182.