ADDED LAGRANGIAN AND SOME OTHER CORRECTIONS
A Lagrangian function has been added to 'functions.py' and appropriately
used now in the tests.

Corrected an error in the docstring of 'kinetic_energy'.

Modified the 'potential_energy' methods for both Particle and RigidBody;
the default value is now zero thus preventing the user from having to
set potential energy when there is none.