Rewrite Diagram from scratch.
In this commit I have radically rewritten the ``Diagram`` class.  The
docstrings are not yet updated here, nor is the test coverage proper.
Further, printing and drawing ``Diagram``'s will not work after this
commit.  I have decided to split those fixes into separate commits,
because this commit is already quite huge.  Almost all docstrings are
bad right now; the TODO remarks will help keep that in mind.

The reason for rewriting ``Diagram`` is as follows.  The original
version did _not_ compute all morphism compositions, as I initially
expected.  Worse even, which morphism compositions it computed was
hash-dependent, and that has backfired very painfully when ``Diagram``
actually started to be used in computing embeddings.  When I fixed the
bug and made it add all compositions, it turned out that ``Diagram`` was
not able to store loop morphisms (morphisms with the same domain and
codomain) properly.  When I fixed that, it turned out that ``Diagram``
would just bail out whenever there was a cycle.  Essentially, this has
made ``Diagram`` totally unusable.

This commit rewrites ``Diagram`` to only store the generator morphisms.
Everything else (like the list of all morphisms) is computed on the fly.
``Diagram`` is now able to know whether it is finite or not, by
detecting whether there are cycles in generators or not.  When asked for
all morphisms, ``Diagram`` gives off a generator now.  If this
``Diagram`` is finite, it will enumerate all morphisms it contains.
Otherwise, it will go on forever.  To make it usable, the ``Diagram``
gives off morphisms in a BFS-like manner.  To see how this may be
exploited, take a look at how ``Diagram.hom`` now works.

This commit makes ``Diagram`` pass all the tests in ``test_baseclasses.py``.
The changes to the tests are rather cosmetic and most often have to do with
changed property names and explicit construction of containers from
generators.  This means that, in the finite case, the new ``Diagram``
behaves essentially like the old ``Diagram``.  One radical change is that
the old version would also add the components of every ``CompositeMorphism``
it received at construction.  The new code does not do that, as this will
most probably severely break many things.  (And I also don't remember why
I had that functionality in the old code in the first place.)