setup.py: Use parenthesis around print statements
In Python 3, the print statement was changed to a function. As such,
running setup.py under Python 3 produces SyntaxErrors. However, for such
simple cases, it is possible to just wrap the entire string with
parenthesis and this will print as expected in all Python versions.
Making this change allows us to print the correct error message (that
use2to3 should be run) if the user tries to use setup.py under Python 3.

This addresses issue 3053.