Diagram: Add a public interface for accessing the strongly connected components.
This commit adds the property ``Diagram.objects_components`` which maps
objects to the corresponding strongly connected components, represented
as ``Diagram``'s, as well the property ``Diagram.condensation`` which
produces the condensation of the diagram (see comments in the code).
This commit also adds the necessary tests and extends the docstrings.