Diagram: Add expanded_generators.
This commit adds the code for enumerating the expanded generators of the
diagram.  There are some intuitive examples of expanded generators in
the docstring of ``_is_expanded_generator``, however, no actual
explanation is provided yet.  The proper explanation will be added
separately, in the commit immediately following the next one.