Diagram: Enforce all generators to be extended generators.
Diagram will now raise a ``ValueError`` if a generator which is not an
extended generator is supplied.  This explicit restriction of the class
of morphisms which can serve as generators is mainly due to how
``Diagram`` handles infinite cases.  In general, it should be possible
to overcome this restriction by carefully handling a number of corner
cases, but, apparently, it's not yet worth it.  Should the need for such
unusual generators arise, it can be implemented.

The next commit will add the documentation for the changes introduced by
this and the previous ones.