Diagram: Add _morphisms_by_stages.
This method essentially represents an internal version of
``Diagram.morphisms`` which, besides yielding the actual morphisms, also
yields the number of the generation round.  This commit also rewrites
``Diagram.morphisms`` to use the new method.

The point behind having this method is that sometimes is very useful to
distinguish between generation rounds in order to be able to check
certain stopping conditions.  See next commit for an example.