add quick_sort for args ordering
When an object has arguments whose order is arbitrary, it is
still a good idea to order them in some canonical way so that
two objects that may have been entered with arguments in different
order will still be the same, e.g. x + 2 == 2 + x. Although hash
sorting is used (in Add, for example) this may run into hash collisions
and so in those cases there should be a way to consistently break ties.
The quick_sort function has been added to iterables to allow for this.
It sorts elements of an iterable on the basis of hash and breaks ties
using the default_sort_key. A tuple is returned from the function.