Polyhedron: cyclic clash; polyhedra have pgroups; Basic.copy
named_groups uses cyclic so the function in permutations was
changed to full_cyclic_form0 and one_based was changed to
cyclic_form1; these are more descriptive and consistent with
the method name (cyclic_form).

All pre-defined polyhedra come with pgroups that are checked
to cover all orientations.

Docstrings have been expanded.

The _canonical function was renamed minlex but it was kept
in the polyhedron file; it is potentially useful for iterables
or geometry.

Unused imports were deleted.

The copy method was added to Basic to allow a copy to be
made of an object that implements in-place changes.