CompositeMorphism: Make flatten return a DerivedMorphism.
Previously, ``CompositeMorphism.flatten`` would return a simple
``NamedMorphism``.  Returning a ``DerivedMorphism`` instead has the
advantages of keeping the link between the two morphisms and, more
importantly, of assuring better comparison semantics.  For example, now
it is possible to distinguish between two flattened morphisms between
the same two objects and with the same name, but which originated from
two different composites.