permutation: clean up DisjointCycle
Thanks to Tom Bachmann for the help in cleaning this up.

Also, dict is used instead of defaultdict since it, too, allows
for a __missing__ method. The printing of the DisjointCycle
as something other than a dict in doctests has not been resolved,
i.e. in doctests

>>> DisjointCycle(1,2)
{1: 2, 2: 1}

while during an interactive session this would give [(1, 2)].