permutation: lmul and other edits; DisjointCycle -> Cycle
Cycle and Permutation both act from R to L in the sense that
given A and B, A*B applies B to the elements and then applies A.

_af_mul works from L to R (but has a reverse option)
lmul accepts a list of permutations and processes them L to R

lmul is used throughout combinatorics to clearly indicate which
order the multiplication is being performed, but it could probably
be removed and the arguments just reversed in order.

A current deficiency with Permutation is that it cannot be rebuilt
from its args when it is in cycle form since the entire cycle is
not stored. A way around this would be to store the entire cycle
but have the print method only show the non-singletons.