remove __mul__ from Cycle
To avoid confusion, only the __call__ method for Cycle is defined while
__mul__ is defined for Permutation:

Permutation([[1, 2]])*Permutation([[2, 3]]) == Cycle(1, 2)(2, 3)

TODO: a possible confusion is that Permutations printed with print_cyclic
= True print as Cycle(1, 2)(2, 3) (for example) and look like a Cycle
but they are actual a Permutation. So the array_form is defined and the
as_list is not. The Cycle prints as something like [(1, 3, 2)].

>>> a = Permutation([[1,2],[5,3]])
>>> a
Cycle(1, 2)(3, 5)
>>> Cycle(a)
[(1, 2), (3, 5)]
>>> Permutation.print_cyclic = False
>>> a
Permutation([0, 2, 1, 5, 4, 3])