_process_morphism: Split up the method.
This method used to be very large due to the amount of heuristics it
incorporated.  This commit essentially factors out the three branches of
the central if-statement of ``_process_morphism`` into three separate
functions.  The four functions resulted after this split up are all less
than two screenful of text, which is easier to both read and maintain.