combinatorics: lmul -> mul; __xor__ -> conjugate
The computation of the product of more than 8 items in _af_mul (now
_af_rmul) was made to only create 1 intermediate list.

The discussion about R or L multiplication was dropped in favor of
just demonstrating what was meant, e.g. a(b(c(i))), and giving a
corresponding doctest.

The __xor__ method was mapped to conjugate so now a^b given a.conjugate(b)

The subgroup_search test for size 16 was xfailed as slow.