combinatorics: Permutations not equal unless size is equal, too
Things are slowed down considerably when testing cyclic forms, and
allowing permutations of different lengths to compare equal leaves
open the possibiliity of subtle errors. Permutations now compare
equal only if their array forms are equal.

An error is also raise for coset_factor if the g to be factored is
not the same length as the permutations in the group.