Try to match against the identity in AssocOp.match.
If we are trying to match a Mul, the code would iterate over its args and test
all possible matching combination. However this misses cases were a Wild would
match on the mutiplicative identity (1). This commit includes the identity (1)
as a possible matching candidate. This fixes issues 674 and 2069.

Note that something like this was already done for Add, but the code used
expr.as_independent instead of the additive identity (0). Of course, this
expr.as_independent would either return the additive constant of the Add, or
simply return zero. The additive constant will already included as part of
expr.make_args. Zero must be added explicitly as a possible match candidate.