perm_groups: minor edits of insert
There is no need to append element [0,...,cmin] to the end of the
range starting from cmin since negative indices handle this just
fine.

ig was changed to gi since it represents g[i] and ig is similar to jg
which is short for JerrumGraph.

There is no need to do double work in testing to see if g is Identity
and then search for the min change. If there is no min change then
it's the identity. The only caveat would be if one got a g like
[2,1,0,3,4,5] and alpha was 3: None would be returned and the routine
would have failed. For safety, an assertion could be made that if i
is None then g[:cmin] == range(cmin)...this has not been done and no
test has failed to it was left out.