Skip statistics doctests
The doctests raise a deprecation warning in Python 2.5 (or higher if they are
explicitly enabled), and there is no easy way to disable them, as they run at
import time, which happens when the doctest module finds the doctests.  Since
the module is deprecated, I don't feel too bad about not doctesting it. At any
rate, as they stand now, the doctests do pass.