Remove the poly_heurisch.py file
This was just a test by me to convert heurisch to use Poly instead of Expr.
The goal was to make it faster, but this didn't work, because the dense
data type that Poly uses makes it painfully slow with the highly multivariate
polynomials that heurisch uses.  If and when Poly becomes faster, we can dig
this up from the git history and try it again, but until then, this is just
cluttering up the filesystem.