3393: _random returns None when eval fails
equals tries randomly evaluating the expression. Without a good way
to know the domain in which to guess, sometimes it guesses in an
invalid domain for which evalf fails. In this case, when a negative
value for n is chosen, evalf fails. A hack is to watch for this
in _random and just bail when this happens.

http://code.google.com/p/sympy/issues/detail?id=3393