matrices: SparseMatrix mat -> smat
[regarding mat -> smat, see below]

minor edits

put a temporary as_int in place until this is merged with 0.7.2

MUTABLEMATRIX

since _cholesky is a parent-class function, make sure it returns
the same type as was received. (SparseMatrix doesn't have a _new
method so either one has to be added or -- as here -- the general
instantiation using the type of self must be used.)

elaborate the _handle_creation_inputs docstring and reorganize
code so there is only one error message location at the end of the
if-blocks.

__setitem__ and __getitem__ are now defined in MutableMatrix and
overridden in SparseMatrix

SPARSEMATRIX

refactored the __init__ code for SparseMatrix and removed the TODO
since it appears that the only things left to process are the
numpy array and the flat list...and I don't see any sparse-smarts
that can be applied to those.

copy is put here to override the MutableMatrix version because...
** smat ** rather than mat is used to store the sparse matrix dict;
if SparseMatrix shares the name 'mat' with MutableMatrix, it is
possible to call methods on Sparse that are only defined to work
in Mutable -- methods that operate on mat. Now, at least, an error
will be raised signaling the need to write a sparse routine for the
desired method.

TODO: change indexing and mat storage so __getitem__ always works
quickly with M[(i, j)] and then perhaps smat can be changed back to
mat.

rowdecomp is such a low-level activity that it is simply inlined
where it is needed.

a duplicate multiply was removed and _lil_row_major removed (and the
row_list is now used in-place in multiply)

all other changes involve changing mat to smat or removing obsolete tests
and editing whitespace.