remove functor from row and col; add row_op/col_op
The dual role of row and col being a selector and in-place modifier
has been split, with the later now being deprecated and replaced with
row_op and col_op, respectively, so DenseMatrix doesn't have a
function capable of changing the matrix. This is in keeping with putting
the mutable-related functions in MutableDenseMatrix. The row_op and
col_op functions were also added to MutableSparseMatrix.