matrices: add test for args
There is a disturbing feature that ImmutableSparseMatrix needs rows and cols
to be defined in the _new method. Here is a record of what happens when
other things are tried:

If I don't assign rows (for example) in the `__init__` then tests fail

```
_______________ sympy/matrices/tests/test_matrices.py:test_cross _______________
  File "/sympy/sympy/matrices/tests/test_matrices.py", line 1971, in test_cross
    test(A.cross(B))
  File "/sympy/sympy/matrices/tests/test_matrices.py", line 1966, in test
    assert ans == M
AssertionError

________________________________________________________________________________
______________ sympy/matrices/tests/test_matrices.py:test_adjoint ______________
  File "/sympy/sympy/matrices/tests/test_matrices.py", line 1996, in test_adjoint
    assert ans == cls(dat).adjoint()
AssertionError
```

And if I put a print statement in the test_cross I get

```
  File "/sympy/sympy/matrices/tests/test_matrices.py", line 1972, in test_cross
    test(A.cross(B))
  File "/sympy/sympy/matrices/tests/test_matrices.py", line 1966, in test
    print type(ans), type(M), M.rows
  File "sympy/matrices/matrices.py", line 2806, in __getattr__
    raise AttributeError("%s has no attribute %s." % (self.__class__.__name__, attr))
AttributeError: ImmutableSparseMatrix has no attribute rows.
```
And that's weird because the `test_args` test (which tests for rows and cols) *passes*.

If I write it as a property like this

```
    @property
    def rows(self):
        return int(self.args[0])
```

I get a complaint from Sparse:

```
  File "sympy/matrices/tests/test_matrices.py", line 1970, in test_cross
    A = cls(a)
  File "sympy/matrices/sparse.py", line 85, in __init__
    self.rows = r
AttributeError: can't set attribute
```

So...I don't understand what is going on. What is there works but it
probably can be done in some better way.