Latex and pretty printing of Mul().
Add brackets if necessary, i.e. when an Integral or a Piecewise is
multiplied by noncommutative terms that cannot be printed in front.

This is needed for Integral, Piecewise, Product, Sum, and possibly
others. Although this is a printing-related issue, it would be nice
if the objects themselves knew if they need brackets or not. For
now, this is implemented in the printers.