test test-suite defs
If a function in the test suite has no argument it is likely a function
with assertions to be tested. But in order to be tested it has to start
with the text 'test_'. A check is now made for this as part of code quality
testing. If one desires to have an arg-less function, they should start
it with an underscore. This requirement makes testing easier in that
one does not have to parse the file to see if the function def has
assertions in it.