use exclude with match variable
Rather than leave the outcome to change, an exclusion is now used
on the pattern symbol so it can't match x.

An interesting thing: if var is used *inside* one of the test
functions, it injects the created variable into the entire
file during testing. I removed that and checked for other
imports/use or var, too.

I touched up some XFAILed tests which couldn't have passed without
the proper variables being defined.