add oset to compatibility
oset is the OrderedSet recipe from ActiveState

http://stackoverflow.com/questions/1653970/does-python-have-an-ordered-set

It is said to be "referred to from the Python Documentation" so it is as
close to being in the stdlib as we might hope for for now.

It keeps track of the order in which items were added to the oset. Items
can be popped from the head or the tail.

Tests were also added to increase the coverage of compatibility.