3462: unification of _eval_power and _eval_expand_power_base
This is something that was sorely needed: much code was being duplicated
(incompletely) between Pow._eval_power and Mul._eval_expand_power_base.
The best of both has now been consolidated under Mul so it is now much
easier to maintain and to see what is part of auto-expansion in Pow and
what is not.

While working on this I realized that powers that weren't strictly Rational
could still allow separation of the bases and made that change in
simplify, too.

args_cnc now has flag to not split -1 from Number (so -2*x has a commutative
part of [-2, x] instead of [-1, 2, x].

The test that was changed in arit was tested at Wolfram Alpha:
http://tinyurl.com/8vsbhzs