Expr._eval_interval is more robust against nan
Previously, it would just check if the result after substitution was nan.  But
nan doesn't always absorb into everything, which would lead to nonsensical
results (from integrate()) containing nan.  An example is issue 1227.

Note that we still do have to use subs, because limit() doesn't always do the
same thing as subs when the limit exists.  For example, limit on Piecewise
raises NotImplementedError.