"Integrate" risch_integrate() into integrate()
Previously, I was going to remove the part of the integrate heuristics that
breaks Integral(g1 + g2 + ...) into Integral(g1) + Integral(g2) + ..., but it
turns out this is a bad idea because many of the heuristics rely on this (for
example, trigintegrate() works on monomials).  Because risch_integrate() works
best on additions that have not been broken up, the strategy I employed was to
call risch_integrate() directly before this loop.  I also call
risch_integrate() again inside the loop, in case the integrand was a sum of an
elementary term and a non-elementary term (like erf(x) + exp(x)).
risch_integrate() is called in the loop after all the heuristics, but before
heurisch(), which should now be considered to be only a last resort.

By default, any NonElementaryIntegral returned from risch_integrate() is
recursively integrated using the other methods, in case it can be written in
terms of special functions.  To use only risch_integrate(), call
integrate(risch=True).  risch_integrate() has been removed from the user
namespace, though it can still be imported manually from
sympy.integrals.risch.  Note that integrate(risch=True) is not equivalent to
risch_integrate() in the sense that the former will return an unevaluated
Integral when it fails and the latter will raise NotImplementedError.
Therefore, you should check isinstance(result, NonElementaryIntegral) if you
want to verify that an unevaluated Integral result is indeed nonelementary.

Many tests had to be rewritten slightly, as risch_integrate() returned answers
in different forms than integrate().  Unfortunately, these are occasionally
worse, notably because risch_integrate() has a habit of returning things like
log(a*b) - log(a) instead of just log(b).

The good news is that for the integrals that risch_integrate() can handle,
integrate() is now orders of magnitude faster.

I also rewrote the grossly outdated and inaccurate docstring of
Integral._eval_integral.