3506: process Rational coeff of exponents
allow processing of sums of powers in _tsolve

watch out for non-log terms in _eval_expand_log (a similar check
is made in the if-block about where the new change was made). The
tests that were added indicated the need for this change here because
a Mul was being generated by log(b) that auto-expanded into a Mul
(for which _eval_expand_log is not defined).