replace broken multiset_partitions
This is a little gem. Perhaps the Knuth routine is better when it's
correct, but this little version (even though it generates all
partitions and the desired one must be selected) is pretty nice.

The table below shows the number of unique partitions obtained
by paritioning n items into 4 groups; the time is also indicated.

n P  time(s)
5 10 0.0
6 65 0.0
7 350 0.01
8 1701 0.03
9 7770 0.11
10 34105 0.62
11 145750 3.82
12 611501 18.28