use ordered in meijerint _mul_as_two_parts
ordered is used in the doctest, not the routine, thus allowing
the calling routine to handle the desired ordering

In `meijerint_indefinite`, ordered is also used with next to
just return the first element of the (perhaps) reordered
results.  It's not necessary to convert the iterator to a list
to do this.