add ibin to iterables; remove arr from shor
From time to time the binary digits of a number are needed.
In string form then can be obtained with bin(n)[2:]. If the integer
form is needed *with a given length* then ibin can be used. ibin
can also be used to give the full sequence of n bits from 0 to 2**n-1
by returning the iterator produced by variations.

The string form of n with length `bits` can also be given.