evalf after as_real_imag; handle _prec=1 in is_comparable
Expressions may give a clear 0 for the imaginary part if they
are allowed to simplify through as_real_imag but if evalf is
done first then a _prec=1 result is obtained:

>>> I**(I+2)).n(2)
-0.21 + 0.e-8*I
>>> _.as_real_imag()[1]
0.e-8
>>> _._prec
1

That "1" means there was no precision computed for the imaginary
part.

>>> (I**(I+2)).as_real_imag()
(-I**I, 0)

Now it's clear to see that the imaginary part is zero.

In is_comparable, the answer was being based on n if there was
an i part that had no precision. It seems better to just return
False in that case. This particular method should return T or F
so if it's not clear that the imaginary part is 0 we'll say that
the expression is not comparable.